chore: move VaultScoresTable and StakedRelayerScoresTable into containers
