+++ src/utils/constants/statuses.ts
@@ -1,0 +1,12 @@
+
+// ray test touch <
+// TODO: could use TypeScript enum
+const STATUSES = Object.freeze({
+  IDLE: 'idle',
+  PENDING: 'pending',
+  RESOLVED: 'resolved',
+  REJECTED: 'rejected'
+});
+
+export default STATUSES;
+// ray test touch >
